Added Nix utilities for developers. Added Gradle for dependency management. Adjusted package scheme.
